转移源相关操作到源编辑页面
