+++ src/components/EditSites.vue
@@ -5,6 +5,15 @@
         <div class="zy-select">
             <div class="vs-placeholder vs-noAfter" @click="openAddSite">添加新源</div>
         </div>
+        <div class="zy-select">
+          <div class="vs-placeholder vs-noAfter" @click="exportSites">导出</div>
+        </div>
+        <div class="zy-select">
+          <div class="vs-placeholder vs-noAfter" @click="importSites">导入</div>
+        </div>
+        <div class="zy-select">
+          <div class="vs-placeholder vs-noAfter" @click="resetSites">重置</div>
+        </div>
         <span class="detail-close zy-svg" @click="close">
           <svg role="img" xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" aria-labelledby="closeIconTitle">
             <title id="closeIconTitle">关闭</title>
@@ -64,8 +73,11 @@
 </template>
 <script>
 import { mapMutations } from 'vuex'
-import { sites } from '../lib/dexie'
+import { sites, setting } from '../lib/dexie'
 import draggable from 'vuedraggable'
+import { remote } from 'electron'
+import { sites as defaultSites } from '../lib/dexie/initData'
+import fs from 'fs'
 export default {
   name: 'editSites',
   data () {
@@ -156,6 +168,62 @@ export default {
         }
         this.$message.success('添加新源成功！')
         this.getSites()
+      })
+    },
+    exportSites () {
+      this.getSites()
+      const arr = [...this.sitesList]
+      const str = JSON.stringify(arr, null, 4)
+      const options = {
+        filters: [
+          { name: 'JSON file', extensions: ['json'] },
+          { name: 'Normal text file', extensions: ['txt'] },
+          { name: 'All types', extensions: ['*'] }
+        ]
+      }
+      remote.dialog.showSaveDialog(options).then(result => {
+        if (!result.canceled) {
+          fs.writeFileSync(result.filePath, str)
+          this.$message.success('已保存成功')
+        }
+      }).catch(err => {
+        this.$message.error(err)
+      })
+    },
+    importSites () {
+      const options = {
+        filters: [
+          { name: 'JSON file', extensions: ['json'] },
+          { name: 'Normal text file', extensions: ['txt'] },
+          { name: 'All types', extensions: ['*'] }
+        ],
+        properties: ['openFile']
+      }
+      remote.dialog.showOpenDialog(options).then(result => {
+        if (!result.canceled) {
+          sites.clear()
+          result.filePaths.forEach(file => {
+            var str = fs.readFileSync(file)
+            const json = JSON.parse(str)
+            sites.bulkAdd(json).then(e => {
+              this.getSites()
+              this.d.site = json[0].key
+              setting.update(this.d).then(res => {
+                this.setting = this.d
+              })
+            })
+            this.$message.success('导入成功')
+          }).catch(err => {
+            this.$message.error(err)
+          })
+        }
+      })
+    },
+    resetSites () {
+      sites.clear()
+      sites.bulkAdd(defaultSites).then(e => {
+        this.getSites()
+        this.$message.success('重置源成功')
       })
     }
   },
